You can play and match all cards now
Matching functionality works with an incorrect edge case. You can
match a third card with a member of an already matched pair.

It should flip over an unmatched pair immediately when you click a third
card rather than making you wait.

Need to add turns and record which card belongs to which player.